Add Notification.timestamp and NotificationOptions.timestamp
Fixes #20.

The `timestamp` of a notification is the time, in milliseconds since
the epoch, of the event for which the notification was created. Web
developers can use this if the time of this event does not match the
time at which the notification is being shown, which can be the case
when notifying the user of upcoming calendar events.