Remove support for sounds from notifications
There are no implementations for this property, and the native
notification centers of various platforms make it hard to support this
properly.

  - Windows Action Center toasts have an <audio> element supports
    built-in sounds, as well as sounds included in an application's
    package, but not freeform path and/or URL support.
  - NSUserNotification (Mac OS and iOS) have a `soundName` property that
    will be resolved against files in various local system directories.
  - Android strongly prefers usage of the RingtoneManager for sounds.

Even though it could be supported on certain platforms, it's currently
underspecified: which codecs should be supported and what is the maximum
duration of the sound (if any), making it very unpredictable at best.

We can revisit supporting sounds if there is interest from
implementations.